pyo3_async_runtimes

Function into_future_with_locals

pub fn into_future_with_locals(
    locals: &TaskLocals,
    awaitable: Bound<'_, PyAny>,
) -> PyResult<impl Future<Output = PyResult<Py<PyAny>>> + Send>

Convert a Python awaitable into a Rust Future

This function converts the awaitable into a Python Task using run_coroutine_threadsafe. A completion handler sends the result of this Task through a futures_channel::oneshot::Sender<PyResult<Py<PyAny>>> and the future returned by this function simply awaits the result through the futures_channel::oneshot::Receiver<PyResult<Py<PyAny>>>.

Arguments

locals - The Python event loop and context to be used for the provided awaitable

awaitable - The Python awaitable to be converted

Examples

use std::time::Duration;
use std::ffi::CString;

use pyo3::prelude::*;

const PYTHON_CODE: &'static str = r#"
import asyncio

async def py_sleep(duration):
    await asyncio.sleep(duration)
"#;

async fn py_sleep(seconds: f32) -> PyResult<()> {
    let test_mod = Python::attach(|py| -> PyResult<Py<PyAny>> {
        Ok(
            PyModule::from_code(
                py,
                &CString::new(PYTHON_CODE).unwrap(),
                &CString::new("test_into_future/test_mod.py").unwrap(),
                &CString::new("test_mod").unwrap(),
            )?
            .into()
        )
    })?;

    Python::attach(|py| {
        pyo3_async_runtimes::into_future_with_locals(
            &pyo3_async_runtimes::tokio::get_current_locals(py)?,
            test_mod
                .call_method1(py, "py_sleep", (seconds,))?
                .into_bound(py),
        )
    })?
    .await?;
    Ok(())
}