pyo3_async_runtimes::testing

Function parse_args

pub fn parse_args() -> Args

Parse the test args from the command line

This should be called at the start of your test harness to give the CLI some control over how our tests are run.

Ideally, we should mirror the default test harness’s arguments exactly, but for the sake of simplicity, only filtering is supported for now. If you want more features, feel free to request them here.

Examples

Running the following function:

let args = parse_args();

Produces the following usage string:

Pyo3 Asyncio Test Suite
USAGE:
test_example [TESTNAME]

FLAGS:
-h, --help       Prints help information
-V, --version    Prints version information

ARGS:
<TESTNAME>    If specified, only run tests containing this string in their names