pyo3_async_runtimes::tokio

Module re_exports

attributes re-exports for macros

Modules

runtime

re-export tokio::runtime to build runtimes in tokio macros without additional dependency The Tokio runtime.

Functions

pending

re-export pending to be used in tokio macros without additional dependency Creates a future which never resolves, representing a computation that never finishes.